chainlit_memory/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
asyncer==0.0.2
attrs==24.3.0
bidict==0.23.1
certifi==2024.12.14
chainlit==1.0.0
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.5.14
Deprecated==1.2.15
exceptiongroup==1.2.2
fastapi==0.100.1
fastapi-socketio==0.0.10
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.12.0
googleapis-common-protos==1.66.0
grpcio==1.69.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.27.0
idna==3.10
importlib_metadata==8.5.0
Lazify==0.4.0
literalai==0.0.1
marshmallow==3.23.3
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.1
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
packaging==23.2
propcache==0.2.1
protobuf==5.29.2
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.10.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-graphql-client==0.4.3
python-multipart==0.0.6
python-socketio==5.12.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.0
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.27.0
syncer==2.0.3
tokenizers==0.21.0
tomli==2.2.1
tqdm==4.67.1
transformers==4.47.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uptrace==1.28.2
urllib3==2.3.0
uvicorn==0.23.2
watchfiles==0.20.0
websockets==14.1
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
